cookiecutter/python/{{cookiecutter.service_name}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.service_name}}"
version = "0.1.0"
description = ""
authors = ["Matias Wallenius <MatiWall@live.dk>"]
readme = "README.md"

[[tool.poetry.source]]
name = "private"
url = "http://pypi.mw.local"
priority = "primary"


[[tool.poetry.source]]
name = "PyPI"
priority = "supplemental"

[tool.poetry.dependencies]
python = "^3.11"
extensions-opentelemetry = "0.1.1"
extensions-configuration = "0.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
